pyproject.toml
[tool.poetry]
name = "learn_poetry"
version = "0.1.0"
description = ""
authors = ["Umar Ali <umeralikhan7864@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
